pyproject.toml
[tool.poetry]
name = "sequence_evolution"
version = "0.1.0"
description = "DNA sequence evolution using Evo2 language model"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11.7"
torch = "^2.0.0"
numpy = "^1.24.0"
biopython = "^1.81"
matplotlib = "^3.7.0"
seaborn = "^0.12.0"
tqdm = "^4.65.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^23.0.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
mypy = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
multi_line_output = 3
